pyproject.toml
[tool.poetry]
name = "voxart"
version = "0.1.0"
description = "Tools for creating voxel art to be 3D printed"
authors = ["Patrick Riley <patriley@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
#python = "^3.9"
python = ">=3.9,<3.11"
numpy = "^1.24.1"
pandas = "^1.5.2"
matplotlib = "^3.6.2"
numpy-stl = "^3.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
jupyter = "^1.0.0"
pytype = "^2022.12.15"
seaborn = "^0.12.2"
pre-commit = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
